% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable.R
\name{summarize_variable}
\alias{summarize_variable}
\title{Summarize Baseball Player Salary or Any Other Variable}
\usage{
summarize_variable(dataset = ISLR::Hitters, variable_name)
}
\arguments{
\item{dataset}{A data frame containing the dataset. Default is the \strong{Hitters} dataset from the ISLR package.}

\item{variable_name}{A string representing the column name of the variable to be summarized (e.g., "Salary", "Hits").}
}
\value{
A list containing:
\item{summary}{Summary statistics for the chosen variable (e.g., Min, 1st Qu., Median, Mean, 3rd Qu., Max).}
\item{mean}{Mean of the chosen variable.}
\item{median}{Median of the chosen variable.}
\item{sd}{Standard deviation of the chosen variable.}
\item{min}{Minimum value of the chosen variable.}
\item{max}{Maximum value of the chosen variable.}
}
\description{
This function provides a summary of any variable (e.g., Salary, Hits) in the dataset,
including basic statistics such as mean, median, standard deviation, and summary statistics.
The function works on the \strong{Hitters} dataset from the ISLR package by default, but can be
applied to any dataset.
}
\examples{
# Load the ISLR package to access the Hitters dataset
library(ISLR)

# Summarize Salary in the Hitters dataset
summarize_variable(Hitters, "Salary")

# Summarize Hits in the Hitters dataset
summarize_variable(Hitters, "Hits")

}
